fix: avoid DropZone leaking style to children
